refactor: layout 하나로 정의
