+++ apps/web/src/app/[locale]/guild/@modal/(.)create/page.tsx
@@ -1,4 +1,4 @@
-import GuildCreate from '../../(components)/GuildCreate';
+import GuildCreate from '../../_components/GuildCreate';
 import GuildModal from '../GuildModal';
 
 export default function GuildCreateModal() {
