+++ apps/web/src/app/[locale]/guild/[id]/setting/GuildSetting.tsx
@@ -6,7 +6,7 @@ import type { Guild } from '@gitanimals/api';
 import { useRouter } from '@/i18n/routing';
 import { updateGuildAction } from '@/serverActions/guild';
 
-import { GuildInfoFormClient, GuildInfoFormSubmitButton } from '../../(components)/GuidlInfoFormClient';
+import { GuildInfoFormClient, GuildInfoFormSubmitButton } from '../../_components/GuidlInfoFormClient';
 
 export interface FormState {
   message: string;
