+++ apps/web/src/app/[locale]/guild/[id]/setting/layout.tsx
@@ -1,25 +1,7 @@
 import type { PropsWithChildren } from 'react';
-import { css } from '_panda/css';
-import { Flex } from '_panda/jsx';
+
+import { GuildModalPageLayout } from '../../_components/GuildModalPageLayout';
 
 export default function GuildCreatePage({ children }: PropsWithChildren) {
-  return (
-    <div className={css({ minH: '100%', display: 'flex', justifyContent: 'center', alignItems: 'center' })}>
-      <div
-        className={css({
-          maxWidth: '880px',
-          mx: 'auto',
-          background: 'gray.gray_150',
-          p: '40px',
-          borderRadius: '16px',
-          color: 'white.white',
-          w: '100%',
-        })}
-      >
-        <Flex flexDirection="column" gap="24px">
-          {children}
-        </Flex>
-      </div>
-    </div>
-  );
+  return <GuildModalPageLayout>{children}</GuildModalPageLayout>;
 }
