+++ apps/web/src/app/[locale]/guild/[id]/setting/page.tsx
@@ -1,7 +1,8 @@
 import { checkIsLeader, getGuildBackgrounds, getGuildById, getGuildIcons } from '@gitanimals/api';
-import { dialogTitleStyle } from '@gitanimals/ui-panda';
 
 import { redirect } from '@/i18n/routing';
+
+import { GuildModalPageTitle } from '../../_components/GuildModalPageLayout';
 
 import { GuildSetting } from './GuildSetting';
 
@@ -18,7 +19,7 @@ export default async function GuildSettingPage({ params }: { params: { id: strin
 
   return (
     <>
-      <h2 className={dialogTitleStyle}>Guild Setting</h2>
+      <GuildModalPageTitle>Guild Setting</GuildModalPageTitle>
       <GuildSetting icons={icons} backgrounds={backgrounds} guildId={params.id} initialData={data} />;
     </>
   );
