+++ apps/web/src/app/[locale]/guild/_components/GuildCard.tsx
