+++ apps/web/src/app/[locale]/guild/_components/GuildDetail.tsx
