+++ apps/web/src/app/[locale]/guild/_components/GuildSearch.tsx
