+++ apps/web/src/app/[locale]/guild/detail/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { GuildDetail } from '../../(components)/GuildDetail';
+import { GuildDetail } from '../../_components/GuildDetail';
 
 export default function GuildDetailPage({ params }: { params: { id: string } }) {
   return <GuildDetail guildId={params.id} />;
